Move all options to options.py, and move options.py to optbinning/options.py
