+++ optbinning/scorecard/counterfactual/counterfactual_information.py
@@ -8,14 +8,7 @@ Counterfactual information.
 from ...binning.binning_information import print_header
 from ...binning.binning_information import print_optional_parameters
 from ...binning.binning_information import print_solver_statistics
-
-
-counterfactual_default_options = {
-    "scorecard": None,
-    "special_missing": False,
-    "n_jobs": 1,
-    "verbose": False
-}
+from ...options import counterfactual_default_options
 
 
 def print_status(status):
